Reorganize macro crate (#1406)
